Split functions into git file around git utilities
